:art: [MODIFY]: seperate bcrypt and jwt related functions to utils folder
